Move upload and export options from 'MenuButton' to Toolbar
