+++ src/components/Header/MenuOptions/ThemeOption.tsx
@@ -8,7 +8,7 @@ type Props = {
 
 const ThemeOption = ({ theme }: Props) => {
     return (
-        <div style= {{ cursor: 'pointer'}}>
+        <div style= {{ cursor: 'pointer', minWidth: 140}}>
             <IconButton color="inherit">
                 {theme === 'dark' ? <Brightness7Icon /> : <Brightness4Icon />}
             </IconButton>
