+++ src/components/Header/ToolbarOptions/FileMenu/FileOption.tsx
@@ -1,7 +1,8 @@
 import { Menu } from "@mui/material"
 import Button from '@mui/material/Button'
 import { useState } from 'react'
-import OpenFile from './FileMenu/OpenFile'
+import ExportFile from './ExportFile'
+import OpenFile from './OpenFile'
 
 type Props = {
   onUpload: (e: React.ChangeEvent<HTMLInputElement>) => void;
@@ -20,10 +21,11 @@ const OpenFileOption = ({ onUpload }: Props) => {
   }
 
   return (
-    <div>
-      <Button onClick={openMenu}>File</Button>
+    <div style = {{display: 'inline-block'}}>
+      <Button style = {{padding: 0}} onClick={openMenu}>File</Button>
       <Menu open={Boolean(anchor)} keepMounted anchorEl={anchor} onClose={closeMenu}>
         <OpenFile onUpload = {onUpload}/>
+        <ExportFile />
       </Menu>
     </div>
   )
